LongTM commit change ticket

+++ o/rate/rate_type.go
@@ -6,10 +6,10 @@ import (
 
 type Rate struct {
 	mongodb.BaseModel `bson:",inline"`
-	TicketIdBk        string `bson:"bticket_id" json:"bticket_id"`
-	RatePoint         int    `bson:"rate_point" json:"rate_point"`
-	Comment           string `bson:"comment" json:"comment"`
-	CustomerId        string `bson:"customer_id" json:"customer_id"`
+	TicketIdBk        string  `bson:"bticket_id" json:"bticket_id"`
+	RatePoint         float32 `bson:"rate_point" json:"rate_point"`
+	Comment           string  `bson:"comment" json:"comment"`
+	CustomerId        string  `bson:"customer_id" json:"customer_id"`
 }
 
 var RateTable = mongodb.NewTable("rate", "rt", 20)
